refactored user block commands PROCERGS#765
